Refactoring_routing
