+++ routers/redis.go
@@ -1,0 +1,9 @@
+package routers
+
+import (
+	"github.com/JuliaKozachuk/BackChat/redisconnect"
+)
+
+func Redis() {
+	redisconnect.ExampleClient()
+}
